clean: Move record ExecutionContextImpl next to interface ExecutionContext
